Rename dev to client to indicate that it is the client side code, update broken references
