+++ client/actions/ChangeSnyppr.js
