+++ client/actions/CurrentFavorites.js
