+++ client/components/FavoriteEntry.jsx
