+++ client/components/FavoriteList.jsx
