+++ client/components/SnypprEntry.jsx
