+++ client/components/SnypprList.jsx
