+++ client/components/SnypprReviewsEntry.jsx
