+++ client/components/SnypprReviewsList.jsx
