+++ client/components/StripeSignUp.jsx
