+++ client/components/TransactionsEntry.jsx
