+++ client/reducers/ChangeSnyppr_reducer.js
