+++ client/reducers/CurrentFavorite_reducer.js
