+++ client/utils/stringConverter.js
