+++ webpack.config.js
@@ -1,6 +1,6 @@
 const path = require('path');
 
-const DEV = path.join(__dirname, '/dev');
+const DEV = path.join(__dirname, '/client');
 const OUTPUT = path.join(__dirname, '/public');
 
 module.exports = {
